Changelog ??/??/24a "??" [stable]
Added TileEntity.cs, TileEntitySaveData.cs
- Added Tile entities, a type of entity with fixed "integer" coordinates

Changelog 21/07/24a "Added Hoe" [stable]

- Added Stone Hoe texture and Item class with basic right click functionality

- Moved RenderDistance from PlayerEntity.cs to GameConfig.cs
- Removed _loadedEntityList from Game1.cs as it wasn't being used (entities belong in their respective world only)
- Changed WorldManager UpdateTile to SetTile() and adjusted functionality and parameters

+++ Rendering/TextureClasses/IndexedTextureAtlas.cs
@@ -33,9 +33,9 @@
             float normalisedOffset = (TextureSize + Padding) / (float)AtlasSize;
             float normalisedTextureSize = TextureSize / (float)AtlasSize;
 
-            texCoords.MinX = (float)column * normalisedOffset;
+            texCoords.MinX = column * normalisedOffset;
             texCoords.MaxX = texCoords.MinX + normalisedTextureSize;
-            texCoords.MinY = (float)row * normalisedOffset;
+            texCoords.MinY = row * normalisedOffset;
             texCoords.MaxY = texCoords.MinY + normalisedTextureSize;
 
             if (UseOffset)
